initial version of run nodes refactor
