+++ packages/celotool/src/e2e-tests/attestations_tests.ts
@@ -1,22 +1,66 @@
 import { ContractKit, newKit } from '@celo/contractkit'
 import { AttestationsWrapper } from '@celo/contractkit/lib/wrappers/Attestations'
 import { assert } from 'chai'
-import { getContext, GethTestConfig, sleep } from './utils'
+import { getContext, sleep } from './utils'
+import { GethRunConfig } from '../lib/geth'
 
 const validatorAddress = '0x47e172f6cfb6c7d01c1574fa3e2be7cc73269d95'
 const phoneNumber = '+15555555555'
 
+const TMP_PATH = '/tmp/e2e'
+
 describe('attestations tests', () => {
-  const gethConfig: GethTestConfig = {
+  const gethConfig: GethRunConfig = {
+    gethRepoPath: '../../../celo-blockchain',
+    runPath: TMP_PATH,
+    genesisPath: TMP_PATH + '/genesis.json',
+    networkId: 1101,
     migrate: true,
-    instances: [
-      { name: 'validator0', validating: true, syncmode: 'full', port: 30303, rpcport: 8545 },
-      { name: 'validator1', validating: true, syncmode: 'full', port: 30305, rpcport: 8547 },
-      { name: 'validator2', validating: true, syncmode: 'full', port: 30307, rpcport: 8549 },
-      { name: 'validator3', validating: true, syncmode: 'full', port: 30309, rpcport: 8551 },
-      { name: 'validator4', validating: true, syncmode: 'full', port: 30311, rpcport: 8553 },
-    ],
+    instances: [],
   }
+
+  gethConfig.instances = [
+    {
+      gethRunConfig: gethConfig,
+      name: 'validator0',
+      validating: true,
+      syncmode: 'full',
+      port: 30303,
+      rpcport: 8545,
+    },
+    {
+      gethRunConfig: gethConfig,
+      name: 'validator1',
+      validating: true,
+      syncmode: 'full',
+      port: 30305,
+      rpcport: 8547,
+    },
+    {
+      gethRunConfig: gethConfig,
+      name: 'validator2',
+      validating: true,
+      syncmode: 'full',
+      port: 30307,
+      rpcport: 8549,
+    },
+    {
+      gethRunConfig: gethConfig,
+      name: 'validator3',
+      validating: true,
+      syncmode: 'full',
+      port: 30309,
+      rpcport: 8551,
+    },
+    {
+      gethRunConfig: gethConfig,
+      name: 'validator4',
+      validating: true,
+      syncmode: 'full',
+      port: 30311,
+      rpcport: 8553,
+    },
+  ]
 
   const context: any = getContext(gethConfig)
   let contractKit: ContractKit
